rename subpath export and test file
