+++ docs/content/2.getting-started/2.usage.md
@@ -102,10 +102,10 @@ Each function, if you hover over it, shows what's going in, and what's coming ou
 You can also call `.toString()` on any input to see the same information at runtime.
 
 ## Type-Level match result (experimental)
-We also provide an experimental feature that allows you to obtain the type-level results of a RegExp match or replace in string literals. To try this feature, please import all helpers from a subpath export `magic-regexp/type-level-regexp` instead of `magic-regexp`.
+We also provide an experimental feature that allows you to obtain the type-level results of a RegExp match or replace in string literals. To try this feature, please import all helpers from a subpath export `magic-regexp/further-magic` instead of `magic-regexp`.
 
 ```ts
-import { createRegExp, exactly, digit } from 'magic-regexp/type-level-regexp'
+import { createRegExp, exactly, digit } from 'magic-regexp/further-magic'
 ```
 
 This feature is especially useful when you want to obtain the type of the matched groups or test if your RegExp matches and captures from a given string as expected.
@@ -123,5 +123,5 @@ myString.match(createRegExp(exactly('foo').or('bar').groupedAs('g1')))
 the type of the matched result will be `null`, or array of union of possible matches `["bar", "bar"] | ["foo", "foo"]` and `result.groups` will be type `{ g1: "bar" } | { g1: "foo" }`.
 
 ::alert
-For more usage details please see the [usage examples](3.examples.md#type-level-regexp-match-and-replace-result-experimental) or [test](https://github.com/danielroe/magic-regexp/blob/main/test/type-level-regexp.test.ts). For type-related issues, please report them to [type-level-regexp](https://github.com/didavid61202/type-level-regexp).
+For more usage details please see the [usage examples](3.examples.md#type-level-regexp-match-and-replace-result-experimental) or [test](https://github.com/danielroe/magic-regexp/blob/main/test/further-magic.test.ts). For type-related issues, please report them to [type-level-regexp](https://github.com/didavid61202/type-level-regexp).
 ::
