+++ docs/content/2.getting-started/3.examples.md
@@ -38,7 +38,7 @@ assert.equal(TENET_RE.test('TEN<==O==>NET'), true)
 ### Type-level RegExp match and replace result (experimental)
 
 ::alert
-This feature is still experimental, to try it please import `createRegExp ` and all `Input` helpers from `magic-regexp/type-level-regexp` instead of `magic-regexp`.
+This feature is still experimental, to try it please import `createRegExp ` and all `Input` helpers from `magic-regexp/further-magic` instead of `magic-regexp`.
 ::
 
 When matching or replacing with literal string such as `magic-regexp v3.2.5.beta.1 just release!`
@@ -51,7 +51,7 @@ import {
   nyOf, 
   digit, 
   wordChar 
-} from 'magic-regexp/type-level-regexp'
+} from 'magic-regexp/further-magic'
 
 const literalString = 'magic-regexp 3.2.5.beta.1 just release!'
