+++ further-magic.d.ts
@@ -1,3 +1,3 @@
 // Legacy stub for previous TS versions
 
-export * from './dist/type-level-regexp'
+export * from './dist/further-magic'
