+++ tsconfig.json
@@ -8,7 +8,7 @@
     "skipLibCheck": true,
     "paths": {
       "magic-regexp/transform": ["./src/transform"],
-      "magic-regexp/type-level-regexp": ["./src/type-level-regexp"],
+      "magic-regexp/further-magic": ["./src/further-magic"],
       "magic-regexp": ["./src/index"]
     }
   },
